+++ src/main/nextflow/co2footprint/Metrics/Converter.groovy
@@ -69,7 +69,12 @@ class Converter {
      * @return Converted quantity with appropriate scale
      */
     static Quantity scaleUnits(double value, String scale='', String unit='', String targetScale=null) {
+        // Define whether byte-sized steps should be assumed
         int scalingFactor = unit == 'B' ? 1024 : 1000
+        // Alternative names for the units
+        final Map<String, String> alternativeScales = ['K': 'k']
+        scale = alternativeScales.containsKey(scale) ? alternativeScales.get(scale) : scale
+        //
         final List<String> scales = ['p', 'n', 'u', 'm', '', 'k', 'M', 'G', 'T', 'P', 'E']  // Units: pico, nano, micro, milli, 0, Kilo, Mega, Giga, Tera, Peta, Exa
         int scaleIndex = getIdx(scale, scales)
 
@@ -123,6 +128,9 @@ class Converter {
      */
     static Quantity scaleTime(Number value, String unit='ms', String targetUnit='s') {
         value = value as BigDecimal
+        // Alternative names for the units
+        final Map<String, String> alternativeUnits = ['m': 'min', 'd': 'days', 'w': 'weeks']
+        unit = alternativeUnits.containsKey(unit) ? alternativeUnits.get(unit) : unit
         // Supported time units
         final List<String> units = ['ns', 'mus', 'ms', 's', 'min', 'h', 'days', 'weeks', 'months', 'years']
         // Conversion factors between units (e.g., 1000 ms = 1 s, 60 s = 1 min, etc.)