+++ src/test/nextflow/co2footprint/CO2FootprintFactoryTest.groovy
@@ -17,17 +17,9 @@
 package nextflow.co2footprint
 
 import nextflow.Session
-import nextflow.executor.NopeExecutor
-import nextflow.processor.TaskHandler
-import nextflow.processor.TaskId
-import nextflow.processor.TaskProcessor
-import nextflow.processor.TaskRun
 import nextflow.trace.TraceObserver
-import nextflow.trace.TraceRecord
 import spock.lang.Specification
 
-import java.nio.file.Files
-import java.nio.file.Path
 /**
  * This class implements various tests.
  *
@@ -50,9 +42,7 @@ class CO2FootprintFactoryTest extends Specification {
 
     def 'check version' () {
         when:
-        CO2FootprintFactory factory = new CO2FootprintFactory()
-        factory.setPluginVersion()
-        String pluginVersion = factory.getPluginVersion()
+        String pluginVersion = CO2FootprintFactory.readPluginVersion()
 
         then:
         pluginVersion == "1.1.0"